mv distance func to tracingfuncs
